wlugoapi/wlugoapi/urls.py
"""wlugoapi URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/2.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path,include
from rest_framework import routers
from wlugoapp import views
router = routers.DefaultRouter()
router.register(r'users', views.UserListView, 'users')
router.register(r'courses', views.CourseListView, 'courses')
router.register(r'ratings', views.RatingListView, 'register')
router.register(r'reviews', views.ReviewListView, 'reviews')
router.register(r'professors', views.ProfessorListView, 'professors')
router.register(r'professor_course', views.Professor_courseListView, 'professor_course')
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenVerifyView,
    TokenRefreshView,
)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include(router.urls)),
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/token/verify/', TokenVerifyView.as_view(), name='token_verify'),
    path('api/rest-auth/registration/', include('rest_auth.registration.urls')),
]

