wlugoapi/requirements.txt
Django==3.2.0
django-cors==0.1
django-crispy-forms==1.9.2
djangorestframework==3.12.4
gunicorn==20.1.0
pytz==2017.3
django-heroku==0.3.1
mysqlclient==2.0.3
django-cors-headers==3.7.0
whitelist==0.2.8
djangorestframework-simplejwt==3.2.2